code cleanup

+++ src/main/java/OCARIoT/RabbitMQ.java
@@ -1,22 +1,24 @@
 package OCARIoT;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
+import com.mongodb.client.model.Filters;
+import com.mongodb.client.model.Updates;
 import org.bson.BsonDocument;
 import org.bson.BsonString;
 import org.bson.Document;
+import org.bson.conversions.Bson;
 import org.json.JSONObject;
 import org.springframework.amqp.core.Message;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.stereotype.Component;
 import java.util.List;
-import java.util.Map;
 import java.util.ResourceBundle;
+import java.util.logging.Logger;
 
 import static com.mongodb.client.model.Filters.eq;
 
@@ -24,6 +26,8 @@ import static com.mongodb.client.model.Filters.eq;
 
 @Component
 public class RabbitMQ {
+
+    private static final Logger LOGGER = Logger.getLogger( RabbitMQ.class.getName() );
 
     private static ResourceBundle rb = ResourceBundle.getBundle("application");
 
@@ -38,65 +42,64 @@ public class RabbitMQ {
 
 
     @RabbitListener(queues = "${rabbitmq.queue.send.notification}")
-    public String notificationService(Message message) throws JsonProcessingException {
+    public void notificationService(Message message) throws JsonProcessingException {
 
-        System.out.println("Recieved Message From RabbitMQ: " + message);
-
+        //LOGGER.log(Level.INFO,"Received message from RabbitMQ "+ message);
 
         byte[] body = message.getBody();
-        String jsonBody = new String(body);
-        ObjectMapper mapper = new ObjectMapper();
-        Map<String, String> map = mapper.readValue(jsonBody, Map.class);
+        JSONObject jsonmsg = new JSONObject(new String(body));
 
-        if (map.containsKey("topic")){
-            String topic = map.get("topic");
-            String title = map.get("title");
-            String content = map.get("body");
+        if (jsonmsg.has("topic")){
+            String topic = (String) jsonmsg.get("topic");
+            String title = (String) jsonmsg.get("title");
+            String content = (String) jsonmsg.get("body");
             try {
                 FirebaseMessage.sendToTopic(topic, title, content);
 
             } catch (FirebaseMessagingException e) {
                 e.printStackTrace();
-                return "Not possible to send message to topic " + topic +".";
+                //return "Not possible to send message to topic " + topic +".";
             }
-            return "Notification sent to topic " + topic+".";
+            //return "Notification sent to topic " + topic+".";
          }
         else{
-            if (map.containsKey("id")){
+            if (jsonmsg.has("id")){
 
-                String title = map.get("title");
-                String content = map.get("body");
-                String userID = map.get("id");
+                String title = (String) jsonmsg.get("title");
+                String content = (String) jsonmsg.get("body");
+                String userID = (String) jsonmsg.get("id");
 
-                long found = collection.countDocuments(new BsonDocument("code", new BsonString(userID)));
+                long found = collection.countDocuments(new BsonDocument("id", new BsonString(userID)));
                 if (found==0){
-
-                    return "User does not exist in database.";
-
+                    //LOGGER.log(Level.INFO,"User does not exist in database ");
+                    //return "User does not exist in database.";
                 }
+                else {
 
-                List<String> tokens = (List) collection.find(eq("id",userID)).first().get("Tokens");
+                    //LOGGER.log(Level.INFO,"Seatching for Tokens");
+                    List<String> tokens = (List) collection.find(eq("id", userID)).first().get("Tokens");
+                    //LOGGER.log(Level.INFO,"Tokens "+ tokens);
+
+                    for (String token : tokens) {
+
+                        try {
+
+                            FirebaseMessage.sendToToken(token, title, content);
 
 
-                for (String token:tokens){
+                        } catch (FirebaseMessagingException e) {
+                            Bson filter = Filters.eq("id",userID);
+                            Bson delete = Updates.pull("Tokens",token);
+                            collection.updateOne(filter,delete);
+                            //e.printStackTrace();
 
-                    try {
-
-                        FirebaseMessage.sendToToken(token, title, content);
-
-
-                    } catch (FirebaseMessagingException e) {
-                        e.printStackTrace();
+                        }
                     }
-
-
+                    //return "Notification message sent to user " + userID;
                 }
-
-               return "Notification message sent to user " + userID;
-
             }
         }
-        return  "Message must contain 'id' or 'topic' keys and values";
+        //return  "Message must contain 'id' or 'topic' keys and values";
     }
 
     @RabbitListener(queues = "${rabbitmq.queue.delete.users}")
@@ -104,15 +107,25 @@ public class RabbitMQ {
 
         byte[] body = message.getBody();
         JSONObject jsonmsg = new JSONObject(new String(body));
-        String id = String.valueOf(jsonmsg.getJSONObject("user").get("id"));
+        //LOGGER.log(Level.INFO,"Received message from RabbitMQ " + message);
+        String id = null;
+
+        if (jsonmsg.has("user")) {
+            id = String.valueOf(jsonmsg.getJSONObject("user").get("id"));
 
 
-        Document doc = collection.find(eq("id",id)).first();
-        if (doc!=null){
+            Document doc = collection.find(eq("id", id)).first();
+            if (doc != null) {
 
-            collection.deleteMany(doc);
+                collection.deleteMany(doc);
+                //LOGGER.log(Level.INFO, "User " + id + " deleted from database");
 
+
+            } else if (doc == null) {
+
+                //LOGGER.log(Level.WARNING, "User " + id + " does not exist on database");
+
+            }
         }
-
     }
 }
